requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0 